cleaning up the event handler code
