+++ frontend/src/components/MeetingView/AgendaView/AgendaView.jsx
@@ -16,7 +16,6 @@ import {
   rectIntersection,
 } from '@dnd-kit/core';
 import {
-  arrayMove,
   sortableKeyboardCoordinates,
 } from '@dnd-kit/sortable';
 
@@ -27,6 +26,7 @@ import MultipleSelectionBox from '../../MultipleSelectionBox/MultipleSelectionBo
 import MeetingItemStates from '../../../constants/MeetingItemStates';
 import { RenderedAgendaItem } from './AgendaItem';
 import { UPDATE_MEETING_ITEM } from '../../../graphql/graphql';
+import { handleDragStart, handleDragOver, handleDragEnd } from './agendaViewFunctions/dndKitFunctions';
 
 /**
  * Used to display a list of a meeting's agenda items and controls to
@@ -226,157 +226,9 @@ function AgendaView({ meeting }) {
   // of which prevents the DND kit from moving items between completed and pending groups
   const displayAgenda = showCompleted ? agendaGroups : createRenderedGroups();
 
-  // These are the event handlers for Dndcontext
-
-  // called when the user starts dragging
-  const handleDragStart = (event) => {
-    const { active } = event;
-
-    setActiveId(active.id);
-  };
-
-  // called when the user drags the dragOverlay on top of a agenda item or the group header
-  // This function will handle the swapping of items between the agenda containers
-  const handleDragOver = (event) => {
-    const { active, over } = event;
-
-    setAgendaGroups((parents) => {
-      const newParents = JSON.parse(JSON.stringify(parents));
-
-      // these are used in the conditional expressions
-      let activeIsOver;
-      let overIsContainer;
-      let activeIsContainer;
-
-      // assigns conditional expressions, has to be put in the if statement in the event the dragged
-      // item is brought out of the draggable zone
-      const overIsNull = over === null;
-      if (!overIsNull) {
-        activeIsOver = active.id === over.id;
-        overIsContainer = parents.filter((parent) => parent.id === over.id).length > 0;
-        activeIsContainer = parents.filter((parent) => parent.id === active.id).length > 0;
-      }
-
-      // item swapping is handled within
-      if (!overIsNull && !activeIsOver && !overIsContainer && !activeIsContainer) {
-        let activeContainerIndex;
-        let overContainerIndex;
-        let activeIndex;
-        let overIndex;
-
-        // finding the values of the variables above
-        newParents.forEach((parent, parentIndex) => {
-          parent.items.forEach((item, itemIndex) => {
-            if (item.id === active.id) {
-              activeIndex = itemIndex;
-              activeContainerIndex = parentIndex;
-            }
-            if (item.id === over.id) {
-              overIndex = itemIndex;
-              overContainerIndex = parentIndex;
-            }
-          });
-        });
-        // entered when the dragOverlay has entered a new agenda group
-        if (activeContainerIndex !== overContainerIndex) {
-          const overIsDropId = newParents.filter((parent) => parent.dropID === over.id).length > 0;
-
-          // This makes sure the selected items are in the correct object containers
-          setSelectedItems(() => {
-            const deepCopy = JSON.parse(JSON.stringify(selectedItems));
-            const activeContainerKey = String(newParents[activeContainerIndex].id);
-            const overContainerKey = String(newParents[overContainerIndex].id);
-            let needToSwap = false;
-
-            const keyIsUndefined = typeof deepCopy[activeContainerKey] === 'undefined';
-
-            if (!keyIsUndefined && deepCopy[activeContainerKey][active.id]) {
-              needToSwap = true;
-              delete deepCopy[activeContainerKey][active.id];
-            }
-
-            // entered only if a swap is needed
-            if (needToSwap) {
-              // entered when no object already assigned, prevents erasing previously checked items
-              // in the agenda container that the dragged item is moving to
-              if (!Object.prototype.hasOwnProperty.call(deepCopy, overContainerKey)) {
-                deepCopy[newParents[overContainerIndex].id] = {};
-              }
-              deepCopy[newParents[overContainerIndex].id][active.id] = true;
-            }
-
-            return deepCopy;
-          });
-
-          // entered when the dragOverlay is not on top of the header
-          if (!overIsDropId) {
-            const itemToMove = newParents[activeContainerIndex].items.splice(activeIndex, 1)[0];
-            itemToMove.parent_meeting_item_id = newParents[overContainerIndex].id;
-            newParents[overContainerIndex].items.splice(overIndex + 1, 0, itemToMove);
-          } else { // entered when the overlay is on top of the header
-            newParents.forEach((parent, i) => {
-              if (parent.dropID === over.id) {
-                overContainerIndex = i;
-              }
-            });
-
-            const itemToMove = newParents[activeContainerIndex].items.splice(activeIndex, 1)[0];
-            itemToMove.parent_meeting_item_id = newParents[overContainerIndex].id;
-            newParents[overContainerIndex].items.push(itemToMove);
-          }
-        }
-      }
-
-      return newParents;
-    });
-  };
-
-  // called when the user lets go of the dragged item
-  const handleDragEnd = (event) => {
-    const { active, over } = event;
-
-    // if statement only entered when the agendaitem is over a valid drop location
-    if (over != null && active.id !== over.id) {
-      setAgendaGroups((parents) => {
-        const newParents = JSON.parse(JSON.stringify(parents));
-
-        let parentIndex; // index of the agendaGroup currently hovered over
-        let oldIndex; // The old index of the agendaItem being moved
-        let newIndex; // The new index of the agendaItem being moved
-
-        parents.forEach((parent, index) => {
-          parent.items.forEach((item, itemIndex) => {
-            if (item.id === active.id) {
-              parentIndex = index;
-              oldIndex = itemIndex;
-            }
-
-            if (item.id === over.id) {
-              newIndex = itemIndex;
-            }
-          });
-        });
-
-        newParents[parentIndex].items = arrayMove(parents[parentIndex].items, oldIndex, newIndex);
-
-        return newParents;
-      });
-    }
-
-    // resort the agendaItems after a drop
-    setAgendaGroups((parents) => {
-      const newParents = JSON.parse(JSON.stringify(parents));
-
-      for (let i = 0; i < newParents.length; i += 1) {
-        for (let j = 0, oN = oNumStart; j < newParents[i].items.length; j += 1, oN += 1) {
-          newParents[i].items[j].order_number = oN;
-        }
-      }
-
-      return newParents;
-    });
-  };
-
+  const onDragStartArgs = { setActiveId };
+  const onDragEndArgs = { setAgendaGroups, oNumStart };
+  const onDragOverArgs = { setAgendaGroups, setSelectedItems, selectedItems };
   return (
     <div className="AgendaView">
       {isAdmin ? <button className="saveOrder" type="button" onClick={saveReOrder}>Save Order</button> : ''}
@@ -395,9 +247,9 @@ function AgendaView({ meeting }) {
       <DndContext
         sensors={sensors}
         collisionDetection={rectIntersection}
-        onDragStart={isAdmin ? handleDragStart : null}
-        onDragEnd={isAdmin ? handleDragEnd : null}
-        onDragOver={isAdmin ? handleDragOver : null}
+        onDragStart={isAdmin ? (e) => { handleDragStart(e, onDragStartArgs); } : null}
+        onDragEnd={isAdmin ? (e) => { handleDragEnd(e, onDragEndArgs); } : null}
+        onDragOver={isAdmin ? (e) => { handleDragOver(e, onDragOverArgs); } : null}
       >
         <Accordion allowZeroExpanded allowMultipleExpanded className="agenda">
           <AgendaGroups
